Update testing framework to modern Chef standards
This is a massive update to bring this cookbook up to modern Chef
standards. It requires ChefDK to be installed and assumes its use,
which is the current best practice versus the Gemfile hell that we
have had to maintain up until now.

- Update Travis CI testing configuration from sous-chef templates
  - Use Chef Delivery Local mode instead of Rake for testing
  - Remove unnecessary files for Rubocop and Foodcritic
  - Get rid of that terrible Gemfile (yay)
  - Use Dokken for testing inside Docker containers
  - Let `cookstyle -a` do it's thing
- Updated ChefSpec platform versions to non-deprecated platforms
  - CentOS - 6.9
  - Ubuntu - 14.04
- Updated ChefSpec tests to have better coverage
  - Use `do_nothing` matchers
- Added a hadoop_test cookbook under test/fixtures/cookbooks
  - HDFS and YARN service setup and start
  - Force `systemctl daemon-reload` on systemd-based systems

The hadoop_test cookbook can be updated to add more test coverage.
This is a first pass and I didn't want to take all the fun. ;-)

Signed-off-by: Chris Gianelloni <wolf31o2@gmail.com>